// src/app/models/firebase-chat.model.ts
export class chatlisting {
  $key: string;
  userId?: string;
  userName?: string;
  userAvatar?: string;
  time?: string;
  message?: string;
  upertext?: string;
  from_id?: string;
  from?: string;
  to_username?: string;
  from_username?: string;
  to_uid?: string;
  from_uid?: string;
  timestamp?: string;
}


export interface FILE {
  name: string;
  filepath: string;
  size: number;
  user: string;
  timestamp: string;
}